Fixes mal-moved Interface
