+++ src/Abstracts/AbstractList.php
@@ -2,13 +2,16 @@
 
 namespace PHPAlchemist\Abstracts;
 
-
 /**
  * Abstract class for List (Roll) Objectified array class
+ *
  * @package PHPAlchemist\Abstracts
  */
 abstract class AbstractList extends AbstractIndexedArray
 {
+    /**
+     * @inheritDoc
+     */
     public function offsetUnset(mixed $offset) : void
     {
         parent::offsetUnset($offset);
@@ -17,6 +20,9 @@ abstract class AbstractList extends AbstractIndexedArray
         }
     }
 
+    /**
+     * @inheritDoc
+     */
     public function offsetset(mixed $offset, mixed $value) : void
     {
         parent::offsetSet($offset, $value);
@@ -25,14 +31,24 @@ abstract class AbstractList extends AbstractIndexedArray
         }
     }
 
+    /**
+     * Reset the keys to 0,1,2,3,...
+     *
+     * @return void
+     */
     protected function rebalanceData() : void
     {
         $this->data = array_values($this->data);
     }
 
+    /**
+     * Determine if array is a List
+     *
+     * @return bool
+     */
     protected function isList() : bool
     {
-        return array_is_list($this->data);
+        return array_is_list($this->data) && ($this instanceof AbstractList);
     }
 
     /**
@@ -42,7 +58,7 @@ abstract class AbstractList extends AbstractIndexedArray
      */
     public function parseData() : mixed
     {
-        foreach ( $this->data as $data) {
+        foreach ($this->data as $data) {
             yield $data;
         }
     }
