+++ src/Contracts/KeyValuePairInterface.php
@@ -2,8 +2,6 @@
 
 namespace PHPAlchemist\Contracts;
 
-use PHPAlchemist\Traits\ArrayInterface;
-
 /**
  * Key Value Pair Interface
  * @package PHPAlchemist\Contracts
